Add todoBtn delete functionality
